🚚 settings_window.rs を役割別の 4 ファイルに分割
- mod.rs（型・tag 変換・再エクスポート）/ rows.rs（AppState 非依存のコントロール生成・配置）/ build.rs（ウィンドウ組み立て）/ sync.rs（AppState と絡む同期・保存・操作）
- 依存は rows → mod / build → rows / sync → rows の一方向（非循環）
- move-only: 関数本体・式は無変更（--color-moved で全行 moved を確認）。外部から見えるパス crate::settings_window::X は再エクスポートで維持し、app.rs の差分ゼロ
- CLAUDE.md のモジュール表・development.md・build_menu_icon.sh のパス参照を追随
