+++ src/settings_window/sync.rs
@@ -1,0 +1,613 @@
+use objc2::msg_send;
+use objc2::runtime::AnyObject;
+
+use crate::app::{
+    apply_settings_now, present_hud, show_sample_image_content, show_text_content, AppState,
+};
+use crate::config::{
+    apply_config_file, default_display_settings, hud_emoji_validation_error, load_config_file,
+    save_config_file, set_config_value, settings_to_config_file, ConfigKey, DisplaySettings,
+    LanguageSetting,
+};
+use crate::i18n::{self, Lang, Msg};
+use crate::objc_helpers::{nsstring_from_str, nsstring_to_string};
+use crate::png::create_preview_sample_image;
+
+use super::rows::{
+    login_item_control_state, select_language_item, select_popup_item, set_string_value,
+};
+use super::{tag_to_config_key, LocalizedKind, SettingsControls};
+
+enum PreviewSample {
+    ShortText,
+    LongText,
+    Image,
+}
+
+const PREVIEW_SAMPLES: [PreviewSample; 3] = [
+    PreviewSample::ShortText,
+    PreviewSample::LongText,
+    PreviewSample::Image,
+];
+
+/// 既定の `max_chars_per_line`・`max_lines` で行内と行数の両方の切り詰めが起きる分量にしてある。
+fn preview_long_text(lang: Lang) -> String {
+    let mut lines = vec![i18n::preview_long_line(lang)];
+    lines.extend((2..=7).map(|n| i18n::preview_numbered_line(lang, n)));
+    lines.join("\n")
+}
+
+/// 言語ポップアップで選択されている設定値を返す。
+unsafe fn selected_language_value(popup: *mut AnyObject) -> Option<LanguageSetting> {
+    if popup.is_null() {
+        return None;
+    }
+    let index: isize = msg_send![popup, indexOfSelectedItem];
+    let index = usize::try_from(index).ok()?;
+    i18n::LANGUAGE_CHOICES.get(index).copied()
+}
+
+/// 言語切り替えに合わせて選択肢のラベルを差し替える。選択位置は動かさない。
+///
+/// `setTitle:` は action を送らないので、この関数を `settingChanged:` の処理中に
+/// 呼んでも再入は起きない。
+unsafe fn retitle_language_popup(popup: *mut AnyObject, lang: Lang) {
+    if popup.is_null() {
+        return;
+    }
+    for (index, setting) in i18n::LANGUAGE_CHOICES.iter().enumerate() {
+        let item: *mut AnyObject = msg_send![popup, itemAtIndex: index as isize];
+        if item.is_null() {
+            continue;
+        }
+        let ns = nsstring_from_str(i18n::language_label(lang, *setting));
+        let () = msg_send![item, setTitle: ns];
+        let () = msg_send![ns, release];
+    }
+}
+
+unsafe fn sync_slider(slider: *mut AnyObject, value_label: *mut AnyObject, value: f64) {
+    let () = msg_send![slider, setDoubleValue: value];
+    set_string_value(value_label, &format!("{value:.2}"));
+}
+
+unsafe fn sync_stepper_field(field: *mut AnyObject, stepper: *mut AnyObject, value: usize) {
+    let () = msg_send![stepper, setIntegerValue: value as isize];
+    set_string_value(field, &value.to_string());
+}
+
+/// 現在の `DisplaySettings` の値をすべてのコントロールに反映する。
+/// ウィンドウを開くたびに呼び、外部（ファイル監視等）での変更との食い違いを防ぐ。
+///
+/// # Safety
+/// - `controls` は `build_settings_window` が返した有効なポインタを保持していること。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn sync_controls_from_settings(controls: &SettingsControls, settings: &DisplaySettings) {
+    sync_slider(
+        controls.poll_interval_slider,
+        controls.poll_interval_value_label,
+        settings.poll_interval_secs,
+    );
+    sync_slider(
+        controls.hud_duration_slider,
+        controls.hud_duration_value_label,
+        settings.hud_duration_secs,
+    );
+    sync_slider(
+        controls.hud_fade_duration_slider,
+        controls.hud_fade_duration_value_label,
+        settings.hud_fade_duration_secs,
+    );
+    sync_slider(
+        controls.hud_scale_slider,
+        controls.hud_scale_value_label,
+        settings.hud_scale,
+    );
+
+    sync_stepper_field(
+        controls.max_chars_per_line_field,
+        controls.max_chars_per_line_stepper,
+        settings.truncate_max_width,
+    );
+    sync_stepper_field(
+        controls.max_lines_field,
+        controls.max_lines_stepper,
+        settings.truncate_max_lines,
+    );
+    sync_stepper_field(
+        controls.hud_image_max_height_field,
+        controls.hud_image_max_height_stepper,
+        settings.hud_image_max_height,
+    );
+
+    select_popup_item(controls.hud_position_popup, settings.hud_position.as_str());
+    select_popup_item(
+        controls.hud_background_color_popup,
+        settings.hud_background_color.as_str(),
+    );
+    select_language_item(controls.language_popup, settings.language);
+
+    set_string_value(controls.hud_emoji_field, &settings.hud_emoji);
+    // プログラムによる同期はすべて妥当な値のみを書き戻すため、残っているメッセージは消す。
+    // controlTextDidChange: 側の再入防止（try_lock）に任せると、ロック保持中の呼び出しでは
+    // 通知が黙って捨てられメッセージが古いまま残りうるため、ここで確実にクリアする。
+    set_string_value(controls.hud_emoji_message_label, "");
+}
+
+/// ログイン項目トグルの表示を実際の LaunchAgent の状態に合わせる。
+/// ウィンドウを開くたびに呼び、ファイル外（Finder 等）での変更との食い違いを防ぐ。
+/// `toggleLoginItem:` が `enable`/`disable` に失敗したときの巻き戻しにも使う。
+///
+/// # Safety
+/// AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn sync_login_item_toggle(controls: &SettingsControls) {
+    let state = login_item_control_state(crate::login_item::is_enabled());
+    let () = msg_send![controls.login_item_toggle, setState: state];
+}
+
+/// 言語ポップアップの選択を、いま効いている設定値に合わせる。
+///
+/// # Safety
+/// AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn sync_language_popup(controls: &SettingsControls, setting: LanguageSetting) {
+    select_language_item(controls.language_popup, setting);
+}
+
+/// `controlTextDidChange:` から呼ぶ。絵文字フィールドの現在の入力値を判定し、
+/// 不正なら理由をメッセージラベルに表示する。妥当なら空にする。
+///
+/// # Safety
+/// - `APP_STATE` のロックは呼び出し側が `try_lock` で取得済みであること。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn update_emoji_validation_message(state: &AppState) {
+    let raw: *mut AnyObject = msg_send![state.settings_controls.hud_emoji_field, stringValue];
+    let text = nsstring_to_string(raw).unwrap_or_default();
+    let lang = i18n::resolve(state.settings.language);
+    let message = hud_emoji_validation_error(&text)
+        .map(|msg| i18n::text(lang, msg))
+        .unwrap_or("");
+    set_string_value(state.settings_controls.hud_emoji_message_label, message);
+}
+
+/// `field`/`stepper` はペアで同じ値を表示する。どちらが `sender` でも、変更後の値を
+/// もう一方に同期しつつ文字列として返す。
+unsafe fn sync_paired_value(
+    sender: *mut AnyObject,
+    field: *mut AnyObject,
+    stepper: *mut AnyObject,
+) -> Option<String> {
+    if sender == field {
+        let raw: *mut AnyObject = msg_send![field, stringValue];
+        let text = nsstring_to_string(raw)?;
+        let parsed: isize = text.trim().parse().ok()?;
+        let () = msg_send![stepper, setIntegerValue: parsed];
+        Some(parsed.to_string())
+    } else {
+        let value: isize = msg_send![stepper, integerValue];
+        let text = value.to_string();
+        set_string_value(field, &text);
+        Some(text)
+    }
+}
+
+unsafe fn raw_value_for_control(
+    key: ConfigKey,
+    sender: *mut AnyObject,
+    controls: &SettingsControls,
+) -> Option<String> {
+    match key {
+        ConfigKey::PollIntervalSecs
+        | ConfigKey::HudDurationSecs
+        | ConfigKey::HudFadeDurationSecs
+        | ConfigKey::HudScale => {
+            let value: f64 = msg_send![sender, doubleValue];
+            // スライダーは連続値を返すため、そのまま文字列化すると 1.0700000000000001 のような
+            // 値が設定ファイルに残る。値ラベルの表示と同じ桁で丸めて揃える。
+            Some(format!("{value:.2}"))
+        }
+        ConfigKey::MaxCharsPerLine => sync_paired_value(
+            sender,
+            controls.max_chars_per_line_field,
+            controls.max_chars_per_line_stepper,
+        ),
+        ConfigKey::MaxLines => {
+            sync_paired_value(sender, controls.max_lines_field, controls.max_lines_stepper)
+        }
+        ConfigKey::HudImageMaxHeight => sync_paired_value(
+            sender,
+            controls.hud_image_max_height_field,
+            controls.hud_image_max_height_stepper,
+        ),
+        ConfigKey::HudPosition | ConfigKey::HudBackgroundColor => {
+            let title: *mut AnyObject = msg_send![sender, titleOfSelectedItem];
+            nsstring_to_string(title)
+        }
+        // 言語は表示ラベルが設定値と別なので、タイトルではなく選択位置から引く。
+        ConfigKey::Language => {
+            selected_language_value(sender).map(|value| value.as_str().to_string())
+        }
+        ConfigKey::HudEmoji => {
+            let value: *mut AnyObject = msg_send![sender, stringValue];
+            nsstring_to_string(value)
+        }
+    }
+}
+
+/// 設定ウィンドウのコントロール変更を下書き（`SettingsControls::draft`）に反映するだけで、
+/// ファイル保存も HUD への適用もしない。確定させるには「保存」（`saveSettings:`）を押す。
+///
+/// クランプ・バリデーションは `set_config_value` に委ねる。
+///
+/// # Safety
+/// - `APP_STATE` をロックしないこと（呼び出し側が既にロックを保持している）。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn apply_setting_change(state: &mut AppState, sender: *mut AnyObject) {
+    let tag: isize = msg_send![sender, tag];
+    let Some(key) = tag_to_config_key(tag) else {
+        return;
+    };
+
+    // 言語は下書き（保存ボタンで確定するモデル）の対象外。選んだ瞬間に設定ファイルへ
+    // 保存し、HUD・設定ウィンドウ・メニューへ即時反映する（自動起動トグルと同じ扱い）。
+    if key == ConfigKey::Language {
+        apply_language_setting_change(state, sender);
+        return;
+    }
+
+    let Some(raw_value) = raw_value_for_control(key, sender, &state.settings_controls) else {
+        return;
+    };
+
+    let mut config = settings_to_config_file(state.settings_controls.draft.clone());
+    let warning = match set_config_value(&mut config, key, &raw_value) {
+        Ok(warning) => warning,
+        Err(error) => {
+            eprintln!("warning: {error}");
+            // 拒否された文字列がフィールドに残ると下書きと食い違うので、下書きの値に戻す。
+            // setStringValue: は action を発火しないため settingChanged: の再入は起きない。
+            if key == ConfigKey::HudEmoji {
+                set_string_value(
+                    state.settings_controls.hud_emoji_field,
+                    &state.settings_controls.draft.hud_emoji,
+                );
+                set_string_value(state.settings_controls.hud_emoji_message_label, "");
+            }
+            return;
+        }
+    };
+    if let Some(warning) = warning {
+        eprintln!("warning: {warning}");
+    }
+
+    state.settings_controls.draft =
+        apply_config_file(state.settings_controls.draft.clone(), &config);
+
+    // クランプ・拒否された場合に備え、実際に反映された値でコントロール表示を揃える。
+    // sender（今まさに編集/操作していたコントロール）自身への setStringValue は避ける:
+    // フィールドエディタが終了処理中に再入すると settingChanged: が再度発火し、
+    // 非再入の APP_STATE ロックでハングしうるため。
+    resync_controls_after_apply(
+        &state.settings_controls,
+        &state.settings_controls.draft,
+        key,
+        sender,
+    );
+}
+
+/// 言語ポップアップの `settingChanged:` から呼ぶ。他の設定と違い下書きを経由せず、選択した
+/// 瞬間に読み込み→検証→保存でファイルへ書き、`apply_settings_now` で HUD・設定ウィンドウ・
+/// メニューへ即時反映する。
+///
+/// `apply_settings_now` の言語差分側で `draft.language` も一緒に同期する。ここで揃えておかないと
+/// 「保存」ボタンが `draft`（開いた時点のスナップショット）を丸ごと書き戻す際に、今しがた
+/// 選んだ言語を古い値で上書きしてしまう。
+///
+/// # Safety
+/// - `APP_STATE` をロックしないこと（呼び出し側が既にロックを保持している）。
+/// - AppKit のメインスレッドから呼ぶこと。
+unsafe fn apply_language_setting_change(state: &mut AppState, sender: *mut AnyObject) {
+    // 表示ラベル（「システムに合わせる」等）は設定値ではないので、選択位置から値を引く。
+    let Some(selected) = selected_language_value(sender) else {
+        return;
+    };
+
+    match save_language_setting(state, selected.as_str()) {
+        Ok(new_settings) => apply_settings_now(state, new_settings),
+        Err(error) => {
+            eprintln!("warning: {error}");
+            // 保存できなかったので、ポップアップの表示を実際に効いている値へ戻す。
+            // 下書きを経由しないコントロールなので、放置すると適用されていない値を
+            // 表示し続ける（ログイン項目トグルの失敗時と同じ扱い）。
+            select_language_item(
+                state.settings_controls.language_popup,
+                state.settings.language,
+            );
+        }
+    }
+}
+
+/// 言語設定を設定ファイルへ保存し、保存後のファイル内容から `DisplaySettings` を組み直す。
+///
+/// 組み直しに `display_settings_from_file` を使うのは、`language` だけ差し替えると
+/// 外部エディタでの変更を取り込まないまま `config_mtime` だけ進んでしまい、
+/// ファイル監視がその変更を二度と拾わなくなるため。
+///
+/// 副作用として、ファイルに保存していない状態は捨てられる。「お試し表示」で下書きを
+/// HUD に当てている最中に言語を切り替えると、HUD はファイルの値に戻る。
+unsafe fn save_language_setting(
+    state: &mut AppState,
+    raw: &str,
+) -> Result<DisplaySettings, String> {
+    let Some(path) = state.config_path.clone() else {
+        return Err("config path is not resolved; cannot save language setting".to_string());
+    };
+    let (mut config, _) = load_config_file(&path).map_err(|error| error.to_string())?;
+    set_config_value(&mut config, ConfigKey::Language, raw).map_err(|error| error.to_string())?;
+    save_config_file(&path, &config).map_err(|error| error.to_string())?;
+    state.config_mtime = std::fs::metadata(&path)
+        .ok()
+        .and_then(|m| m.modified().ok());
+
+    crate::app::display_settings_from_file(&path).map_err(|error| error.to_string())
+}
+
+/// 言語切り替えで、設定ウィンドウ内の静的なラベル・タイトルをすべて差し替える。
+///
+/// `APP_STATE` のロックを保持したまま呼んでよい: ここで触るのは非編集の `NSTextField`
+/// ラベルと `NSButton`/`NSWindow` のタイトルだけで、いずれも action やテキスト編集の
+/// デリゲート通知（`controlTextDidChange:` 等）を発火しないため、Objective-C 側からの
+/// 再入は起きない。
+///
+/// # Safety
+/// AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn apply_language(controls: &SettingsControls, lang: Lang) {
+    for entry in &controls.localized {
+        let text = i18n::text(lang, entry.msg);
+        match entry.kind {
+            LocalizedKind::StringValue => set_string_value(entry.control, text),
+            LocalizedKind::Title => {
+                let ns = nsstring_from_str(text);
+                let () = msg_send![entry.control, setTitle: ns];
+                let () = msg_send![ns, release];
+            }
+            LocalizedKind::TabLabel => {
+                let ns = nsstring_from_str(text);
+                let () = msg_send![entry.control, setLabel: ns];
+                let () = msg_send![ns, release];
+            }
+        }
+    }
+    // 「システムに合わせる」は表示言語で書き分けるため、選択肢のラベルも差し替える。
+    retitle_language_popup(controls.language_popup, lang);
+}
+
+/// 下書きを既定値に戻し、コントロール表示を同期する。保存も HUD への適用もしない。
+/// 言語は下書きモデルの対象外のため、既定値へは戻さず現在値を保つ。
+///
+/// # Safety
+/// - `APP_STATE` をロックしないこと（呼び出し側が既にロックを保持している）。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn reset_settings(state: &mut AppState) {
+    let language = state.settings_controls.draft.language;
+    state.settings_controls.draft = default_display_settings();
+    state.settings_controls.draft.language = language;
+    sync_controls_from_settings(&state.settings_controls, &state.settings_controls.draft);
+}
+
+/// 下書きを HUD に適用したうえで固定サンプルを表示する。クリップボードの内容には依存せず、
+/// 押すたびに短文 → 長文 → 画像 → 短文 … と巡回する。設定の見た目を毎回同じ内容で
+/// 比較できるようにするため、クリップボードは参照しない。ファイル保存はしない。
+///
+/// # Safety
+/// - `APP_STATE` をロックしないこと（呼び出し側が既にロックを保持している）。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn preview_settings(this: &AnyObject, state: &mut AppState) {
+    sync_draft_from_controls(&mut state.settings_controls);
+
+    let draft = state.settings_controls.draft.clone();
+    apply_settings_now(state, draft);
+
+    let index = state.settings_controls.preview_sample_index;
+    state.settings_controls.preview_sample_index = (index + 1) % PREVIEW_SAMPLES.len();
+
+    let lang = i18n::resolve(state.settings.language);
+    match PREVIEW_SAMPLES[index] {
+        PreviewSample::ShortText => {
+            show_text_content(state, i18n::text(lang, Msg::PreviewShortText))
+        }
+        PreviewSample::LongText => show_text_content(state, &preview_long_text(lang)),
+        PreviewSample::Image => match create_preview_sample_image() {
+            Ok(image) => show_sample_image_content(state, image),
+            Err(error) => {
+                eprintln!("warning: {error}");
+                show_text_content(state, i18n::text(lang, Msg::PreviewShortText));
+            }
+        },
+    }
+
+    present_hud(this, state);
+}
+
+/// 下書きを設定ファイルへ保存し、HUD に適用する。`config_mtime` を保存後のファイルの mtime に
+/// 更新し、直後のファイル監視ポーリングによる二重適用を防ぐ。
+///
+/// # Safety
+/// - `APP_STATE` をロックしないこと（呼び出し側が既にロックを保持している）。
+/// - AppKit のメインスレッドから呼ぶこと。
+pub unsafe fn save_settings(state: &mut AppState) -> bool {
+    sync_draft_from_controls(&mut state.settings_controls);
+
+    let Some(path) = state.config_path.clone() else {
+        eprintln!("warning: config path is not resolved; cannot save settings");
+        return false;
+    };
+
+    let config = settings_to_config_file(state.settings_controls.draft.clone());
+    if let Err(error) = save_config_file(&path, &config) {
+        eprintln!("warning: {error}");
+        return false;
+    }
+
+    let draft = state.settings_controls.draft.clone();
+    apply_settings_now(state, draft);
+    state.config_mtime = std::fs::metadata(&path)
+        .ok()
+        .and_then(|m| m.modified().ok());
+    true
+}
+
+/// 全キーのコントロールの現在値から下書きを組み直す。
+///
+/// NSTextField の action はフォーカスが外れるか Enter を押すまで飛ばず、その Enter は
+/// デフォルトボタン（保存）に先取りされる。保存・お試し表示の直前にここで読み直すことで、
+/// 未確定の入力の取りこぼしと、拒否された値がフィールドに残ることの両方を防ぐ。
+unsafe fn sync_draft_from_controls(controls: &mut SettingsControls) {
+    const KEYS: [ConfigKey; 10] = [
+        ConfigKey::PollIntervalSecs,
+        ConfigKey::HudDurationSecs,
+        ConfigKey::HudFadeDurationSecs,
+        ConfigKey::HudScale,
+        ConfigKey::MaxCharsPerLine,
+        ConfigKey::MaxLines,
+        ConfigKey::HudImageMaxHeight,
+        ConfigKey::HudPosition,
+        ConfigKey::HudBackgroundColor,
+        ConfigKey::HudEmoji,
+    ];
+
+    let mut config = settings_to_config_file(controls.draft.clone());
+    for key in KEYS {
+        let control = control_for_key(controls, key);
+        if control.is_null() {
+            continue;
+        }
+        let Some(raw) = raw_value_for_key(key, control) else {
+            continue;
+        };
+        if let Err(error) = set_config_value(&mut config, key, &raw) {
+            eprintln!("warning: {error}");
+        }
+    }
+    controls.draft = apply_config_file(controls.draft.clone(), &config);
+
+    let draft = controls.draft.clone();
+    sync_controls_from_settings(controls, &draft);
+}
+
+/// キーに対応する「値の入力元」となるコントロール。field と stepper の組は、利用者が直接
+/// 打ち込める field を正とする。
+fn control_for_key(controls: &SettingsControls, key: ConfigKey) -> *mut AnyObject {
+    match key {
+        ConfigKey::PollIntervalSecs => controls.poll_interval_slider,
+        ConfigKey::HudDurationSecs => controls.hud_duration_slider,
+        ConfigKey::HudFadeDurationSecs => controls.hud_fade_duration_slider,
+        ConfigKey::HudScale => controls.hud_scale_slider,
+        ConfigKey::MaxCharsPerLine => controls.max_chars_per_line_field,
+        ConfigKey::MaxLines => controls.max_lines_field,
+        ConfigKey::HudImageMaxHeight => controls.hud_image_max_height_field,
+        ConfigKey::HudPosition => controls.hud_position_popup,
+        ConfigKey::HudBackgroundColor => controls.hud_background_color_popup,
+        ConfigKey::HudEmoji => controls.hud_emoji_field,
+        ConfigKey::Language => controls.language_popup,
+    }
+}
+
+unsafe fn raw_value_for_key(key: ConfigKey, control: *mut AnyObject) -> Option<String> {
+    match key {
+        ConfigKey::PollIntervalSecs
+        | ConfigKey::HudDurationSecs
+        | ConfigKey::HudFadeDurationSecs
+        | ConfigKey::HudScale => {
+            let value: f64 = msg_send![control, doubleValue];
+            Some(format!("{value:.2}"))
+        }
+        ConfigKey::HudPosition | ConfigKey::HudBackgroundColor => {
+            let title: *mut AnyObject = msg_send![control, titleOfSelectedItem];
+            nsstring_to_string(title)
+        }
+        // 言語は表示ラベルが設定値と別なので、タイトルではなく選択位置から引く。
+        ConfigKey::Language => {
+            selected_language_value(control).map(|value| value.as_str().to_string())
+        }
+        ConfigKey::MaxCharsPerLine
+        | ConfigKey::MaxLines
+        | ConfigKey::HudImageMaxHeight
+        | ConfigKey::HudEmoji => {
+            let value: *mut AnyObject = msg_send![control, stringValue];
+            nsstring_to_string(value)
+        }
+    }
+}
+
+unsafe fn resync_controls_after_apply(
+    controls: &SettingsControls,
+    settings: &DisplaySettings,
+    key: ConfigKey,
+    sender: *mut AnyObject,
+) {
+    match key {
+        ConfigKey::PollIntervalSecs => {
+            set_string_value(
+                controls.poll_interval_value_label,
+                &format!("{:.2}", settings.poll_interval_secs),
+            );
+        }
+        ConfigKey::HudDurationSecs => {
+            set_string_value(
+                controls.hud_duration_value_label,
+                &format!("{:.2}", settings.hud_duration_secs),
+            );
+        }
+        ConfigKey::HudFadeDurationSecs => {
+            set_string_value(
+                controls.hud_fade_duration_value_label,
+                &format!("{:.2}", settings.hud_fade_duration_secs),
+            );
+        }
+        ConfigKey::HudScale => {
+            set_string_value(
+                controls.hud_scale_value_label,
+                &format!("{:.2}", settings.hud_scale),
+            );
+        }
+        ConfigKey::MaxCharsPerLine => resync_stepper_field(
+            sender,
+            controls.max_chars_per_line_field,
+            controls.max_chars_per_line_stepper,
+            settings.truncate_max_width,
+        ),
+        ConfigKey::MaxLines => resync_stepper_field(
+            sender,
+            controls.max_lines_field,
+            controls.max_lines_stepper,
+            settings.truncate_max_lines,
+        ),
+        ConfigKey::HudImageMaxHeight => resync_stepper_field(
+            sender,
+            controls.hud_image_max_height_field,
+            controls.hud_image_max_height_stepper,
+            settings.hud_image_max_height,
+        ),
+        // ポップアップの選択肢は常に妥当な値のみを取りうるためクランプが発生せず、再同期は不要
+        ConfigKey::HudPosition | ConfigKey::HudBackgroundColor => {}
+        // 拒否された入力は set_config_value のエラー側で戻すため、ここでは触らない。
+        // 編集中フィールドへの setStringValue を避ける狙いもある。
+        ConfigKey::HudEmoji => {}
+        // 言語は下書きモデルの対象外で、保存の成否にかかわらず再同期する対象が無い。
+        ConfigKey::Language => {}
+    }
+}
+
+/// stepper は常に更新する（テキスト編集の再入リスクがないため）。
+/// field は `sender` 自身でない場合のみ更新する（sender の場合は編集中の可能性があるため触れない）。
+unsafe fn resync_stepper_field(
+    sender: *mut AnyObject,
+    field: *mut AnyObject,
+    stepper: *mut AnyObject,
+    value: usize,
+) {
+    let () = msg_send![stepper, setIntegerValue: value as isize];
+    if sender != field {
+        set_string_value(field, &value.to_string());
+    }
+}
